+++ src/applications/vaos/express-care/components/ExpressCareRequestLimitPage.jsx
@@ -1,10 +1,10 @@
 import React, { useEffect } from 'react';
 import { connect } from 'react-redux';
 import AlertBox from '@department-of-veterans-affairs/formation-react/AlertBox';
-import { FETCH_STATUS } from '../utils/constants';
+import { FETCH_STATUS } from '../../utils/constants';
 import { Link } from 'react-router-dom';
-import { selectExpressCareNewRequest } from '../utils/selectors';
-import { scrollAndFocus } from '../utils/scrollAndFocus';
+import { selectExpressCareNewRequest } from '../../utils/selectors';
+import { scrollAndFocus } from '../../utils/scrollAndFocus';
 
 function ExpressCareRequestLimitPage({
   fetchRequestLimitsStatus,
