+++ src/applications/vaos/tests/new-express-care-request/select-reason-page.unit.spec.jsx
@@ -10,7 +10,7 @@ import {
   renderWithStoreAndRouter,
 } from '../mocks/setup';
 import { setupExpressCareMocks } from '../mocks/helpers';
-import ExpressCareReasonPage from '../../containers/ExpressCareReasonPage';
+import ExpressCareReasonPage from '../../express-care/components/ExpressCareReasonPage';
 
 const initialState = {
   user: {
